Prepare CGAffineTransformation interface. Replace Float types with Double. Prettify all around.

+++ Sources/Tokenizer/Layout/Constraint.swift
@@ -100,8 +100,8 @@ public struct Constraint {
 }
 
 public enum ConstraintType {
-    case constant(Float)
-    case targeted(target: ConstraintTarget, targetAnchor: LayoutAnchor, multiplier: Float, constant: Float)
+    case constant(Double)
+    case targeted(target: ConstraintTarget, targetAnchor: LayoutAnchor, multiplier: Double, constant: Double)
 }
 
 extension ConstraintType: Equatable {
